wip split adapters to own files
